Made 'handleSignal' public.
This has to be done, because if a command is running some code outside the context of the command class, a Exception will be triggered in Case of a signal. The cause is that the handleSignal is private, and cant be accessed from code outside the command.
Only and simplest way to deal with that, is to simply make that method public.